Large speedups via vectorization for allnan
```
$ asv compare HEAD^ HEAD -s --sort ratio

Benchmarks that have improved:

       before           after         ratio
     [a72e0d23]       [a8565469]
     <allnan~1>       <allnan>  
-        779±10μs         707±10μs     0.91  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     3.85±0.09μs      3.49±0.04μs     0.91  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 0, 'fast') [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-         776±7μs          696±3μs     0.90  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-      1.14±0.4ms          894±6μs     0.78  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.02±0.01ms         796±30μs     0.78  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.04±0.02ms          774±3μs     0.74  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-         678±8μs          258±6μs     0.38  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-         692±6μs          261±7μs     0.38  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-         927±6μs          258±4μs     0.28  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-         926±7μs          255±3μs     0.28  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', None, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
```